# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GitHub Migrator API",
    "description": "API for managing repository discovery, profiling, batch organization, and migration execution between GitHub instances.",
    "version": "1.2.0",
    "contact": {
      "name": "GitHub Migrator"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v1/repositories": {
      "get": {
        "tags": ["repositories"],
        "summary": "List repositories",
        "description": "List all repositories with optional filtering and pagination",
        "operationId": "listRepositories",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by migration status (comma-separated for multiple)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "batch_id",
            "in": "query",
            "description": "Filter by batch ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by source system",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "organization",
            "in": "query",
            "description": "Filter by organization (comma-separated for multiple)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "query",
            "description": "Filter by ADO project (comma-separated for multiple)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "has_lfs",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_submodules",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_large_files",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_actions",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_wiki",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_pages",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_discussions",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_projects",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_packages",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_environments",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_secrets",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_variables",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_webhooks",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_branch_protections",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_rulesets",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_code_scanning",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_dependabot",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_secret_scanning",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "has_codeowners",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "is_archived",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "is_fork",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "visibility",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["public", "private", "internal"]
            }
          },
          {
            "name": "min_size",
            "in": "query",
            "description": "Minimum repository size in bytes",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "max_size",
            "in": "query",
            "description": "Maximum repository size in bytes",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "size_category",
            "in": "query",
            "description": "Filter by size category (comma-separated)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "complexity",
            "in": "query",
            "description": "Filter by complexity (comma-separated)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search repositories by name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "available_for_batch",
            "in": "query",
            "description": "Filter repositories available for batch assignment",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of repositories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "repositories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Repository"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repositories/{fullName}": {
      "get": {
        "tags": ["repositories"],
        "summary": "Get repository details",
        "description": "Get detailed information about a specific repository including migration history",
        "operationId": "getRepository",
        "parameters": [
          {
            "$ref": "#/components/parameters/fullName"
          }
        ],
        "responses": {
          "200": {
            "description": "Repository details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "repository": {
                      "$ref": "#/components/schemas/Repository"
                    },
                    "history": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MigrationHistory"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["repositories"],
        "summary": "Update repository",
        "description": "Update repository metadata such as batch assignment and priority",
        "operationId": "updateRepository",
        "parameters": [
          {
            "$ref": "#/components/parameters/fullName"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "batch_id": {
                    "type": "integer",
                    "format": "int64",
                    "nullable": true
                  },
                  "priority": {
                    "type": "integer"
                  },
                  "destination_full_name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repository updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Repository"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/batches": {
      "get": {
        "tags": ["batches"],
        "summary": "List batches",
        "description": "Get all migration batches",
        "operationId": "listBatches",
        "responses": {
          "200": {
            "description": "List of batches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Batch"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["batches"],
        "summary": "Create batch",
        "description": "Create a new migration batch",
        "operationId": "createBatch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Batch created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/batches/{id}": {
      "get": {
        "tags": ["batches"],
        "summary": "Get batch details",
        "description": "Get detailed information about a specific batch",
        "operationId": "getBatch",
        "parameters": [
          {
            "$ref": "#/components/parameters/batchId"
          }
        ],
        "responses": {
          "200": {
            "description": "Batch details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "batch": {
                      "$ref": "#/components/schemas/Batch"
                    },
                    "repositories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Repository"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["batches"],
        "summary": "Update batch",
        "description": "Update batch metadata",
        "operationId": "updateBatch",
        "parameters": [
          {
            "$ref": "#/components/parameters/batchId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "scheduled_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["batches"],
        "summary": "Delete batch",
        "description": "Delete a batch",
        "operationId": "deleteBatch",
        "parameters": [
          {
            "$ref": "#/components/parameters/batchId"
          }
        ],
        "responses": {
          "200": {
            "description": "Batch deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/batches/{id}/repositories": {
      "post": {
        "tags": ["batches"],
        "summary": "Add repositories to batch",
        "description": "Add repositories to a batch",
        "operationId": "addRepositoriesToBatch",
        "parameters": [
          {
            "$ref": "#/components/parameters/batchId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["repository_ids"],
                "properties": {
                  "repository_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repositories added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "batch": {
                      "$ref": "#/components/schemas/Batch"
                    },
                    "repositories_added": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["batches"],
        "summary": "Remove repositories from batch",
        "description": "Remove repositories from a batch",
        "operationId": "removeRepositoriesFromBatch",
        "parameters": [
          {
            "$ref": "#/components/parameters/batchId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["repository_ids"],
                "properties": {
                  "repository_ids": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "format": "int64"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Repositories removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "batch": {
                      "$ref": "#/components/schemas/Batch"
                    },
                    "repositories_removed": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Repository": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "full_name": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "source_url": {
            "type": "string"
          },
          "total_size": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "has_lfs": {
            "type": "boolean"
          },
          "has_submodules": {
            "type": "boolean"
          },
          "has_large_files": {
            "type": "boolean"
          },
          "has_actions": {
            "type": "boolean"
          },
          "has_wiki": {
            "type": "boolean"
          },
          "has_pages": {
            "type": "boolean"
          },
          "has_discussions": {
            "type": "boolean"
          },
          "has_projects": {
            "type": "boolean"
          },
          "is_archived": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "batch_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "priority": {
            "type": "integer"
          },
          "complexity_score": {
            "type": "integer"
          },
          "destination_url": {
            "type": "string",
            "nullable": true
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "migrated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "MigrationHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "repository_id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_message": {
            "type": "string",
            "nullable": true
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "duration_seconds": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    },
    "parameters": {
      "fullName": {
        "name": "fullName",
        "in": "path",
        "required": true,
        "description": "Repository full name (org/repo)",
        "schema": {
          "type": "string"
        }
      },
      "batchId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Batch ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}